+++ bindings/python/capstone/arm.py
@@ -53,7 +53,7 @@ class ArmOp(ctypes.Structure):
         ('type', ctypes.c_uint),
         ('value', ArmOpValue),
         ('subtracted', ctypes.c_bool),
-        ('access', ctypes.c_uint8),
+        ('access', ctypes.c_uint),
         ('neon_lane', ctypes.c_int8),
     )
